api/requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
click==8.2.1
colorama==0.4.6
crispy-bootstrap5==2025.4
Django==5.2.1
django-crispy-forms==2.4
et_xmlfile==2.0.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
joblib==1.5.2
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlparse==0.5.3
starlette==0.47.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
